[IMPROVEMENTS] Update how permutations work
