+++ backend/storage/tools/scan.rb
@@ -14,7 +14,6 @@ require_relative 'libs/tools/amass'
 require_relative 'libs/tools/c99'
 require_relative 'libs/tools/dehashed'
 require_relative 'libs/tools/gau'
-require_relative 'libs/tools/gotator'
 require_relative 'libs/tools/gowitness'
 require_relative 'libs/tools/httpx'
 require_relative 'libs/tools/mesh'
